update landing page design and create component for user guide
